Keep Go AUTO integer fields aligned with literals

+++ tests/integration/cases/record_wide_int/Go_heterogeneous_strategy_record_suffix_auto@v1_18.go
@@ -1,6 +1,6 @@
 package main
 type Record0 struct {
-	Quantity int
+	Quantity int64
 	Big uint64
 	Ratio float64
 	Label string
